Support proper generation of TimeGenerated field
The log collector API expects that the TimeGenerated field is passed in
a field NOT named "TimeGenerated" but identified in a special
http-header.

The prometheus provider used to just create a "TimeGenerated" entry per
log object which was just ignored by log analytics. This has been changed
to "TimeGeneratedPrometheus" and self.colTimeGenerated is set accordingly.